refactor : Transectional 어노테이션 사용 정리

+++ src/main/java/roomescape/time/service/ReservationTimeService.java
@@ -12,11 +12,11 @@ import roomescape.time.repository.ReservationTimeRepository;
 
 @Service
 @RequiredArgsConstructor
-@Transactional
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
 
+    @Transactional
     public ReservationTimeResponse save(ReservationTimeRequest reservationTimeRequest) {
         LocalTime startAt = reservationTimeRequest.startAt();
         ReservationTime reservationTime = ReservationTime.createNew(startAt);
@@ -34,7 +34,11 @@ public class ReservationTimeService {
                 .toList();
     }
 
+    @Transactional
     public void deleteById(long id) {
+        if(!reservationTimeRepository.existsById(id)){
+            throw new IllegalArgumentException("삭제할 예약 시간이 존재하지 않습니다.");
+        }
         reservationTimeRepository.deleteById(id);
     }
